requirements.txt
gradio
python-dotenv
sentence-transformers
faiss-cpu
pymupdf
requests
